requirements.txt
requests
pandas
pyarrow
mysql.connector==8.3.0
python-dotenv
sqlalchemy